+++ app/src/components/core/empty/EmptyStateGraphic.tsx
