+++ app/src/components/core/empty/LinkCard.tsx
