+++ app/src/components/index.tsx
@@ -25,8 +25,6 @@ export * from "./core/LoadingMask";
 export * from "./core/ViewSummaryAside";
 export * from "./core/copy";
 export * from "./core/SectionHeading";
-export * from "./core/Empty";
-export * from "./empty-state";
 export * from "./exception";
 export * from "./core/KeyboardToken";
 export * from "./color/ColorSwatch";
